Replace SocketManager with api.js - the start of a big overhaul
