+++ src/components/App.js
@@ -1,19 +1,39 @@
-import React from "react";
+import React, { useState } from "react";
 import GlobalStyles from "../styles/globalStyles";
-import SocketManager from "./SocketManager";
 import Header from "./Header";
 import Main from "./Main";
 import Game from "./Game";
+import { doEmitter } from "../api";
 
 function App() {
+    const [roomName, setRoomName] = useState("");
+
+    const startRoom = () => {
+        doEmitter("event/start-room", (acknowledgement) => {
+            if (!acknowledgement.ok) {
+                console.log(`Failed to start room "${acknowledgement.roomName}".`);
+            } else {
+                setRoomName(acknowledgement.roomName);
+            }
+        });
+    };
+
+    const joinRoom = (roomName) => {
+        doEmitter("event/join-room", roomName, (acknowledgement) => {
+            if (!acknowledgement.ok) {
+                console.log(`Failed to join room "${acknowledgement.roomName}".`);
+            } else {
+                setRoomName(acknowledgement.roomName);
+            }
+        });
+    };
+
     return (
         <div className="App">
             <GlobalStyles />
-            <SocketManager>
-                <Header />
-                <Main />
-                <Game />
-            </SocketManager>
+            <Header roomName={roomName} />
+            <Main startRoom={startRoom} joinRoom={joinRoom} />
+            <Game />
         </div>
     );
 }
